refactor(action): consolidate type definitions into single module
- Delete separate action-record.ts and action.ts files
- Create new type.ts module exporting Action and ActionRecord interfaces
- Update all imports in delegate.ts, method.ts, and main.ts to use type.ts
- Rename lib.ts to lib_.ts for consistency with registry_.ts naming
- Rename registry.ts to registry_.js for consistency with internal module pattern
- Simplify import paths and reduce module fragmentation
- Maintain full backward compatibility of public API
