+++ pkg/nanolib/action/src/registry_.ts
@@ -1,59 +1,4 @@
-import type {Action} from './action.js';
-
-// ─── Type Definitions ────────────────────────────────────────────────────────
-
-/**
- * A modifier handler used in `on-<eventType>` attribute syntax.
- *
- * Receives the triggering DOM `event`, the `element` that owns the
- * `on-<eventType>` attribute, and the **mutable** `action` object being built.
- * The handler may mutate `action.meta` to attach cross-cutting data (e.g. a
- * trace ID, a timestamp, or an A/B flag) before the action reaches subscribers.
- *
- * Return `true` (or any truthy value) to allow the action to proceed, or
- * `false` to cancel the dispatch entirely.
- *
- * Using explicit parameters instead of `this` binding makes handlers
- * compatible with arrow functions and easier to test in isolation.
- *
- * @example — a modifier that stamps a timestamp into meta
- * ```ts
- * const timestampHandler: ModifierHandler = (_event, _element, action) => {
- *   action.meta ??= {};
- *   action.meta['timestamp'] = Date.now();
- *   return true;
- * };
- * ```
- *
- * @example — a modifier that cancels dispatch when the element is disabled
- * ```ts
- * const notDisabledHandler: ModifierHandler = (_event, element) => {
- *   return !(element as HTMLButtonElement).disabled;
- * };
- * ```
- */
-export type ModifierHandler = (event: Event, element: HTMLElement, action: Action) => boolean;
-
-/**
- * A payload resolver used in `on-<eventType>` attribute syntax.
- *
- * Receives the triggering DOM `event` and the `element` that owns the
- * `on-<eventType>` attribute. The return value becomes the `payload` field of
- * the `Action` object passed to `onAction` subscribers.
- *
- * Using explicit parameters instead of `this` binding makes resolvers
- * compatible with arrow functions and easier to test in isolation.
- *
- * @example — a resolver that returns the element's dataset id
- * ```ts
- * const dataIdResolver: PayloadResolver = (_event, element) => {
- *   return (element as HTMLElement).dataset.id ?? null;
- * };
- * ```
- */
-export type PayloadResolver = (event: Event, element: HTMLElement) => unknown;
-
-// ─── Registries ──────────────────────────────────────────────────────────────
+import type {ModifierHandler, PayloadResolver} from './type.js';
 
 /**
  * Registry of all named modifier handlers.
